Moved EnvironmentVar class to the model package
